[algospot/SFW] solved / cpp / 16ms

+++ algospot/accepted/SFW.cc
@@ -7,7 +7,7 @@ int main() {
 	int T;
 	cin >> T;
 	while(T--) {
-		int N,M,r0,c0,r,c,A,B;
+		long long N,M,r0,c0,r,c,A,B;
 		cin >> N >> M >> r0 >> c0 >> r >> c >> A >> B;
 		cout << (abs(r-r0)+(A-1))/A + (abs(c-c0)+(B-1))/B << endl;
 	}
